Refactoring ViewPort Part 3
Some of these methods belonged as extension methods on IDataModel. They're only used in the ViewPort, but don't really do anything ViewPort specific... they act on a Model, with a few parameters. So make them called that way.
